fix(Clause): match whole variable when inlining using interpolate
Previously when inlining variables, a simple string replacement was used which meant that shorter
variables could be inlined into longer ones. Eg. a variable "$param" could be inlined into the
string "$paramName" to produce 'value'Name, which is obviously incorrect. The string replacement
operation now attempts to only inline a param if it matches the entire variable.