fix(Where): bind where clause parameters during build
Previously, the combined where clause was created in the constructor and then reused later. This
meant that the parameter bag wouldn't be able to create a unique name for parameters in where
comparison functions because the where clause hadn't been attached to a query yet. This commit
changes this behaviour so that the where clause is created during the build call.