fix(Set): replace override option with merge
The override option has now been replaced with the merge option to be more consistent with cypher.
The merge option is just the opposite of the override option. Setting to true will cause the `+=`
operator to be used in the query.

This also affects the default behaviour of the Set class and all `.set*` family functions. There
was previously an inconsistency between the Set class and the set methods on the builder interface
in regards to whether to use the `=` or `+=` operator by default which has now been normalized to
always use `=`. This means that the new merge option will always default to false.

BREAKING CHANGE: The default behaviour of the Set clause has changed to use the `=` operator.
This is to be more consistent with cypher.