fix(Set): change default value of the override option to false
Previously, if no options were provided to the Set class then it would set override to true. This
has now changed to false to fall in line with the default values for methods on the builder class.
Usages of the set method on the builder class remain unchanged in their behaviour.

BREAKING CHANGE: The default value of the override option has changed from true to false when
directly instantiating the Set class.