feat(create): add unique option
The create clause now accepts an additional options parameter that contains a `unique` flag. Setting
it to true will cause the create clause to only create the row if it doesn't already exist. In
addition, the `createUnique` and `createUniqueNode` methods were added to the query interface as
shortcuts for `create(..., { unique: true })` and `createNode(..., ..., ..., { unique: true })`
respectively.

fix #105